Move Until, Forever, NeverStop, ForeverTestTimeout from util to wait
