+++ pkg/util/util_test.go
@@ -18,39 +18,7 @@ package util
 
 import (
 	"testing"
-	"time"
 )
-
-func TestUntil(t *testing.T) {
-	ch := make(chan struct{})
-	close(ch)
-	Until(func() {
-		t.Fatal("should not have been invoked")
-	}, 0, ch)
-
-	ch = make(chan struct{})
-	called := make(chan struct{})
-	go func() {
-		Until(func() {
-			called <- struct{}{}
-		}, 0, ch)
-		close(called)
-	}()
-	<-called
-	close(ch)
-	<-called
-}
-
-func TestUntilReturnsImmediately(t *testing.T) {
-	now := time.Now()
-	ch := make(chan struct{})
-	Until(func() {
-		close(ch)
-	}, 30*time.Second, ch)
-	if now.Add(25 * time.Second).Before(time.Now()) {
-		t.Errorf("Until did not return immediately when the stop chan was closed inside the func")
-	}
-}
 
 func TestStringDiff(t *testing.T) {
 	diff := StringDiff("aaabb", "aaacc")
